+++ src/presets.ts
@@ -29,6 +29,7 @@ const highPreset: ObfuscateOptions = {
 
   calculator: true,
   compact: true,
+  hexadecimalNumbers: false,
   controlFlowFlattening: 0.75,
   deadCode: 0.2,
   dispatcher: true,
@@ -64,6 +65,7 @@ const mediumPreset: ObfuscateOptions = {
 
   calculator: true,
   compact: true,
+  hexadecimalNumbers: false,
   controlFlowFlattening: 0.5,
   deadCode: 0.025,
   dispatcher: 0.75,
@@ -92,6 +94,7 @@ const lowPreset: ObfuscateOptions = {
 
   calculator: true,
   compact: true,
+  hexadecimalNumbers: false,
   controlFlowFlattening: 0.25,
   deadCode: 0.01,
   dispatcher: 0.5,