+++ CHANGELOG.md
@@ -1,8 +1,20 @@
+<!-- # `1.5.0`
+Hexadecimal Numbers
+
+- The hexadecimal representation can now be used. [#29](https://github.com/MichaelXF/js-confuser/issues/29)
+- New option `hexadecimalNumbers` to control this behavior.
+
+### `hexadecimalNumbers`
+
+Uses the hexadecimal representation (`50` -> `0x32`) for numbers. (`true/false`) -->
+
 # `1.4.3`
 Minify Fix
 
 - Fixed [#34](https://github.com/MichaelXF/js-confuser/issues/34)
 - - Minify was removing `return` statements too aggressively, fixed in this version.
+
+- Slight improvement to Control Flow Flattening
 
 # `1.4.2`
 Eval Fix